Arcade: Pac-Man and Ms. Pac-Man
floooh's chips vendored (zlib, so it ships with the project unlike the SNES
core): the Z80, its memory mapper, and the Namco board. Ms. Pac-Man is the
bootleg that ran on an unmodified Pac-Man board, so there is no daughterboard
to emulate -- it is just another six-ROM set.

An arcade board's ROMs are several chips and the game list is one row per file,
so tools/mkarcade.py packs a romset into a single .arc on the machine that
already has it, and arcrom parses the container (host-tested, like nesrom and
gbrom). Merged romsets keep clones in subfolders and store each file once, so
the packer matches on name wherever in the zip it sits. No ROMs are included.

The cabinet's monitor is on its side, so the raster is turned upright while the
palette is applied -- the same pass either way. emu_video now takes a
fractional magnification, which is what lets a 288-line picture fill the
panel's height exactly at 2.5x.

It runs at half speed, near 30 fps. The emulator is cycle-stepped, ticking the
CPU and the sound chip three million times a second: about 30 ms of work for
each 16.7 ms of machine time. Internal RAM made no difference and neither did
IRAM -- the cost is the work, not where it lives. Driving it by elapsed time
made it worse, running away to a twelfth of real speed, so a fixed frame of
machine time is what it does. Full speed needs an instruction-stepped CPU.

+++ src/emu_video.h
@@ -24,9 +24,12 @@ struct Geometry {
 // Safe to call repeatedly; false means neither is available.
 bool begin();
 
-// Sets the source size and magnification, and allocates the frames. Returns
-// false if there is no memory for them.
-bool configure(int src_w, int src_h, int scale);
+// Sets the source size and magnification, and allocates the frames. The
+// magnification may be fractional — the hardware scaler takes it either way,
+// and 2.5x is what makes a 288-line arcade picture exactly fill the panel.
+// Returns false if there is no memory for the frames, or if a fractional
+// magnification is asked for on a machine that has no hardware scaler.
+bool configure(int src_w, int src_h, float scale);
 
 // Frees the frames. Waits for any transfer still in flight first.
 void release();
